dom_cfg: add Linux domain with pv drivers
Add Linux domain which with pvnet and pvblk frontends.
This Domain is Linux image with initrd, provided as an external
sources. It is configured with pvnet and pvblock predefined
configuration.
PVNET is configured with ip address 192.168.0.2/24 so it can be
accessible via network from DomD.
PVBLK is configured to take /etc/u-boot-initial-env as an image file
which was taken from the DomD rootfs.

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Reviewed-by: Grygorii Strashko <grygorii_strashko@epam.com>
Acked-by: Dmytro Firsov <dmytro_firsov@epam.com>
Acked-by: Dmytro Semenets <dmytro_semenets@epam.com>